FEAT: create profile modal window
