+++ my-app/src/css/MainPage.css
@@ -132,6 +132,7 @@
 .chat-member-list li {
   border-bottom: 1px solid #3c3c3c;
   padding-left: 5px;
+  display: flex;
 }
 
 .chat-member-list li input {
@@ -167,7 +168,6 @@
   top: 15%;
 }
 
-
 .drawer-side .button-side {
   display: inline-block;
   text-align: center;
@@ -196,16 +196,15 @@
 }
 
 .drawer-side .list {
-	width: 100%;
-	height: 90%;
-	margin-bottom: 30px;
-	overflow: auto;
+  width: 100%;
+  height: 90%;
+  margin-bottom: 30px;
+  overflow: auto;
 }
 
 .drawer-side .menu {
-	min-height: 1200px;
+  min-height: 1200px;
 }
-
 
 .drawer-side .search-side {
   width: 100%;
