+++ my-app/src/css/Modal.css
@@ -1,33 +1,3 @@
-.profile-modal-back {
-  background-color: rgb(128, 4, 4);
-  width: 10000px;
-  height: 10000px;
-  position: fixed;
-  /* position: fixed; */
-  /* width: 100vw;
-  width: 100vh; */
-}
-
-.modal-overlay-back {
-  width: 300%;
-  height: 300%;
-  max-width: 1000px;
-  max-height: 1000px;
-  /* z 포지션을 최상위로  */
-  z-index: 1000;
-
-  position: fixed;
-  /* position: absolute; */
-  top: 50%;
-  /* left: 10%; */
-  right: 10%;
-  transform: translate(-50%, -50%);
-
-  background-color: gray;
-  border: 1px solid black;
-  color: black;
-  border-radius: 8px;
-}
 .profile-modal {
   width: 300%;
   height: 300%;
@@ -37,9 +7,8 @@
   z-index: 1000;
 
   position: fixed;
-  /* position: absolute; */
   top: 50%;
-  /* left: 10%; */
+
   right: 10%;
   transform: translate(-50%, -50%);
