+++ my-app/src/page/MainPage.tsx
@@ -5,6 +5,7 @@ import GameWaiting from "../component/GameWaiting";
 import LeaderBoard from "../component/LeaderBoard";
 import FriendsList from "../component/FriendsList";
 import ChannelsList from "../component/ChannelsList";
+import ProfileModal from "../component/ProfileModal";
 
 export default function MainPage() {
   const [curPage, setCurPage] = useState("leaderboard");
@@ -117,15 +118,15 @@ export default function MainPage() {
               <ul>
                 <li>
                   <input type="checkbox" />
-                  daechoi
+                  <ProfileModal name="daechoi"></ProfileModal>
                 </li>
                 <li>
                   <input type="checkbox" />
-                  dhwlddj
+                  <ProfileModal name="youhan"></ProfileModal>
                 </li>
                 <li>
                   <input type="checkbox" />
-                  qnfekfr
+                  <ProfileModal name="gyyu"></ProfileModal>
                 </li>
               </ul>
               <div className="chat-member-button">
@@ -149,23 +150,21 @@ export default function MainPage() {
             className="menu p-4 w-80 h-full bg-base-200 text-base-content"
             style={{ color: "#8a8a8a" }}
           >
-			<div className="button-side">
-			<button
-				className={friendsButtonClass}
-				onClick={() => handleButtonClick("friends_list")}
-			>
-				FRIENDS
-			</button>
-			<button
-				className={channelsButtonClass}
-				onClick={() => handleButtonClick("channels_list")}
-			>
-				CHANNELS
-			</button>
-			</div>
-			<div className='list'>
-            	{renderSide()}
-			</div>
+            <div className="button-side">
+              <button
+                className={friendsButtonClass}
+                onClick={() => handleButtonClick("friends_list")}
+              >
+                FRIENDS
+              </button>
+              <button
+                className={channelsButtonClass}
+                onClick={() => handleButtonClick("channels_list")}
+              >
+                CHANNELS
+              </button>
+            </div>
+            <div className="list">{renderSide()}</div>
             <div className="search-side">
               <input type="text"></input>
               <button>🔍</button>
